+++ tests/test_aa_warmup.c
@@ -17,8 +17,8 @@ int test(LIBSSH2_SESSION *session)
     (void)hostkey;
 
     fprintf(stdout,
-            "libssh2_session_hostkey returned len, type: %d, %d\n",
-            (int)len, type);
+            "libssh2_session_hostkey returned len, type: %ld, %d\n",
+            (long)len, type);
 
     return 0;  /* always return success */
 }
